renderer_vulkan: split up blit screen resources into separate antialias and window adapt passes
